rename eth_expect to expect_eth
